Ny tickcounter

+++ DTO/MålingDTO.cs
@@ -9,7 +9,7 @@ namespace DTO
 {
     public class MålingDTO
     {
-        public int Puls { get; set; }
+        public long Puls { get; set; }
         public int MiddelBT { get; set; }
         public int Dia { get; set; }
         public int Sys { get; set; }
